# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Behavioral Grounding Framework API
  version: 1.0.0
  summary: REST API for the BGF synthetic-society simulation engine.
  description: |
    HTTP surface served by `api/app.py` (Flask). Covers liveness probes,
    simulation launch, status polling, agent interviews, anchor (omniscient)
    queries, AI-designed scenario generation, the human-evaluation pipeline,
    and the cross-model benchmark spec.

    **Auth model.** `GET`/`HEAD`/`OPTIONS` are always public. All
    `POST`/`PUT`/`DELETE` require `Authorization: Bearer <BGF_API_TOKEN>`
    when the environment variable is set. A small set of write endpoints
    is exempted (see `_PUBLIC_POST_PATHS` in `api/app.py`).

    **Source of truth.** This spec is hand-maintained alongside the Flask
    routes. Drift is caught by `make openapi-lint` and the
    `tests/test_openapi_spec.py` route-coverage check.
  contact:
    name: BGF Maintainer
    url: https://github.com/lucastourinho/SyntheticSocieties
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://touro3-synthetic-societies.hf.space
    description: HuggingFace Space (production)
  - url: http://localhost:5050
    description: Local Flask dev server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Set via `BGF_API_TOKEN` env var on the server.
  parameters:
    ExperimentId:
      name: exp_id
      in: path
      required: true
      schema:
        type: string
      description: Experiment directory name under `experiments/`.
...
  schemas:
...
    RunStatus:
      type: object
      properties:
        experiment_id: { type: string }
        status: { type: string, enum: [running, complete, failed] }
        total_rounds: { type: integer }
        completed_rounds: { type: integer }
        started_at: { type: number }
        updated_at: { type: number }
        finished_at: { type: number, nullable: true }
        error_message: { type: string, nullable: true }
        stale: { type: boolean }
        gpu_wait: { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: { type: string, nullable: true }
paths:
...
  /status/{exp_id}:
    get:
      tags: [simulation]
      summary: Poll progress + freshness for a running or completed run.
      security: []
      parameters:
        - $ref: "#/components/parameters/ExperimentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...